backend/requirements.txt
fastapi==0.115.8
uvicorn==0.15.0
python-dotenv==0.19.0
websockets==10.0
pydantic==2.10.6
pydantic_core==2.27.2
aiohttp==3.9.1
setuptools==70.0.0
